Factorize simple custom action commands
