+++ internal/commands/render.go
@@ -4,8 +4,6 @@ import (
 	"io"
 	"os"
 
-	"github.com/deislabs/duffle/pkg/action"
-	"github.com/deislabs/duffle/pkg/claim"
 	"github.com/docker/app/internal"
 	"github.com/docker/cli/cli"
 	"github.com/docker/cli/cli/command"
@@ -43,36 +41,11 @@ func renderCmd(dockerCli command.Cli) *cobra.Command {
 }
 
 func runRender(dockerCli command.Cli, appname string, opts renderOptions) error {
-	defer muteDockerCli(dockerCli)()
-
-	c, err := claim.New("render")
+	a, c, err := prepareCustomAction("render", dockerCli, appname, opts.registryOptions, opts.pullOptions, opts.parametersOptions)
 	if err != nil {
 		return err
 	}
-	driverImpl, err := prepareDriver(dockerCli, bindMount{})
-	if err != nil {
-		return err
-	}
-	bundle, err := resolveBundle(dockerCli, appname, opts.pull, opts.insecureRegistries)
-	if err != nil {
-		return err
-	}
-	c.Bundle = bundle
-
-	parameters, err := mergeBundleParameters(c.Bundle,
-		withFileParameters(opts.parametersFiles),
-		withCommandLineParameters(opts.overrides),
-	)
-	if err != nil {
-		return err
-	}
-	c.Parameters = parameters
 	c.Parameters[internal.Namespace+"render-format"] = opts.formatDriver
-
-	a := &action.RunCustom{
-		Action: internal.Namespace + "render",
-		Driver: driverImpl,
-	}
 
 	var writer io.Writer = dockerCli.Out()
 	if opts.renderOutput != "-" {
